+++ src/restSpa/db.nim
@@ -2,7 +2,7 @@ import pkg/norm/sqlite
 export sqlite
 
 import std/locks
-export locks
+# export locks
 
 var dbLock*: Lock
 initLock dbLock
@@ -10,6 +10,7 @@ initLock dbLock
 var dbConn* {.guard: dbLock.}: DbConn
 
 template inDb*(body: untyped) =
+  ## This is a simple template to edit db using locks to be `gcsafe`
   {.gcsafe.}:
     withLock dbLock:
       body